refactor: replace Rollup with Rolldown for the build
