+++ .gitattributes
@@ -1,0 +1,3 @@
+setup/dist/*.cjs linguist-generated=true
+report/dist/*.cjs linguist-generated=true
+run/dist/*.cjs linguist-generated=true
