+++ package.json
@@ -20,7 +20,8 @@
   },
   "type": "module",
   "scripts": {
-    "build": "rollup -c"
+    "build": "rolldown -c",
+    "typecheck": "tsc"
   },
   "dependencies": {
     "@sigstore/bundle": "^5.0.0",
@@ -28,12 +29,9 @@
     "@sigstore/verify": "^4.0.0"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^29.0.0",
-    "@rollup/plugin-json": "^6.0.0",
-    "@rollup/plugin-node-resolve": "^16.0.0",
-    "@rollup/plugin-replace": "^6.0.3",
-    "@rollup/plugin-terser": "^1.0.0",
+    "@types/node": "^24.9.2",
     "fast-check": "^4.8.0",
-    "rollup": "^4.60.4"
+    "rolldown": "^1.1.5",
+    "typescript": "^7.0.2"
   }
 }
